remove report build and default test from caliper-flow
